+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 #Changelog
+
 All Notable changes to `League\Period` will be documented in this file
 
 ## 3.X.X - XXXX-XX-XX
@@ -11,6 +12,8 @@ All Notable changes to `League\Period` will be documented in this file
 
 - `Period` always returns `DateTimeImmutable` objects
 - `Period::split` returns a `Generator`
+- `Period` now takes into account microseconds
+- `Period::merge` no longer throws `RuntimeException`
 
 ### Deprecated